refactor: game 패키지 생성 및 일부 클래스 이동
game 패키지 생성 및 일부 클래스 이동

- 로또 게임 운영에 대한 책임이 있는 클래스 -> domain/game 패키지 하위로 이동
