+++ src/main/java/lotto/controller/RaceController.java
@@ -4,12 +4,12 @@ import java.util.List;
 import java.util.stream.Collectors;
 import lotto.model.domain.BonusNumber;
 import lotto.model.domain.Lotto;
-import lotto.model.domain.LottoMachine;
+import lotto.model.domain.game.LottoMachine;
 import lotto.model.domain.Lottos;
 import lotto.model.domain.PurchaseAmount;
-import lotto.model.domain.Rank;
+import lotto.model.domain.game.Rank;
 import lotto.model.domain.WinningNumber;
-import lotto.model.domain.WinningStatistics;
+import lotto.model.domain.game.WinningStatistics;
 import lotto.util.InputParser;
 import lotto.util.InputValidator;
 import lotto.view.InputView;
